Players can now jump up multiple levels
So I change the jumpingOnto/climbingOnto logic for the players so that they have to firstly be in a particular range so that the animation looks reasonable. Additionally jumping/climbing onto objects is default disabled unless the object is tagged "KnightJumpOnto" or "RobotJumpOnto", this is so the players can't climb everythin/ can't climb things of incorrect height. (Knight climbs thing of height 1 and robot climbs things of height 2)