Pushing works
Pushing works for robot. 'Pushable' script needs to be attached to the object which you want the characters to be able to push. RobotRigidbody (useGravity and isKinematic unchecked). The cube/object to be pushed should use gravity and be super heavy so that it doesn't move when walked into, but only when pushed.