Skeleton Prefab
Could not apply root motion so had to make movement controller. In your AI script you just need to set the transform.forward for direction and whichever bools in the animator controller. can't get knockBack animation to look smooth. If you set knockBack my advice is to immediately set it back to false otherwise it plays and plays and plays.