Lever
Ok so I made a "lever" script, if a player is within distance and pushes the assigned pullLever key the lever will be set to activated if it isn't and activated=false if it was activated. Chase can you please make a lever with an appropriate movement of the stick if it is activated vs not. You just need to add the 'lever' script to the prefab.